+++ packages/jest-environment-jsdom/src/__tests__/jsdom_environment.test.ts
@@ -39,4 +39,28 @@ describe('JSDomEnvironment', () => {
 
     expect(env.dom.window.navigator.userAgent).toEqual('foo');
   });
+
+  /**
+   * When used in conjunction with Custom Elements (part of the WebComponents standard)
+   * setting the global.document to null too early is problematic because:
+   *
+   * CustomElement's disconnectedCallback method is called when a custom element
+   * is removed from the DOM. The disconnectedCallback could need the document
+   * in order to remove some listener for example.
+   *
+   * global.close calls jsdom's Window.js.close which does this._document.body.innerHTML = "".
+   * The custom element will be removed from the DOM at this point, therefore disconnectedCallback
+   * will be called, so please make sure the global.document is still available at this point.
+   */
+  it('should not set the global.document to null too early', () => {
+    const env = new JSDomEnvironment(makeProjectConfig());
+
+    const originalCloseFn = env.global.close.bind(env.global);
+    env.global.close = () => {
+      originalCloseFn();
+      expect(env.global.document).not.toBeNull();
+    };
+
+    return env.teardown();
+  });
 });